don't init sentry outside prod; some router refactoring
