+++ backend/api/completions.ts
@@ -1,4 +1,6 @@
 import { Request, Response } from "express"
+import { chunk } from "lodash"
+import * as yup from "yup"
 
 import {
   Completion,
@@ -18,10 +20,26 @@ import { ApiContext } from "./"
 
 const JSONStream = require("JSONStream")
 
-export function completions(ctx: ApiContext) {
-  return async function (req: any, res: any) {
-    const { knex } = ctx
-    const organizationResult = await getOrganization(ctx)(req, res)
+const languageMap: Record<string, string> = {
+  en: "en_US",
+  sv: "sv_SE",
+  fi: "fi_FI",
+}
+
+interface RegisterCompletionInput {
+  completion_id: string
+  student_number: string
+  eligible_for_ects?: boolean
+  tier?: number
+  registration_date?: string
+}
+export class CompletionController {
+  constructor(readonly ctx: ApiContext) {}
+
+  completions = async (req: Request<{ slug: string }>, res: Response) => {
+    const { slug } = req.params
+    const { knex } = this.ctx
+    const organizationResult = await getOrganization(this.ctx)(req, res)
 
     if (organizationResult.isErr()) {
       return organizationResult.error
@@ -35,7 +53,7 @@ export function completions(ctx: ApiContext) {
       await knex
         .select<any, Course[]>("*")
         .from("course")
-        .where({ slug: req.params.course })
+        .where({ slug })
         .limit(1)
     )?.[0]
 
@@ -44,7 +62,7 @@ export function completions(ctx: ApiContext) {
         await knex
           .select("course_id")
           .from("course_alias")
-          .where({ course_code: req.params.course })
+          .where({ course_code: slug })
       )[0]
       if (!course_alias) {
         return res.status(404).json({ message: "Course not found" })
@@ -80,33 +98,20 @@ export function completions(ctx: ApiContext) {
     const stream = sql.stream().pipe(JSONStream.stringify()).pipe(res)
     req.on("close", stream.end.bind(stream))
   }
-}
 
-export function completionInstructions({ knex }: ApiContext) {
-  return async function (req: any, res: any) {
-    const id = req.params.id
-    let language = req.params.language
+  completionInstructions = async (
+    req: Request<{ slug: string; language: string }>,
+    res: Response,
+  ) => {
+    const { knex } = this.ctx
+    const { slug, language } = req.params
 
     const course = (
-      await knex.select<any, Course[]>("id").from("course").where("slug", id)
+      await knex.select<any, Course[]>("id").from("course").where("slug", slug)
     )[0]
 
     if (!course) {
       return res.status(404).json("")
-    }
-
-    switch (language) {
-      case "en":
-        language = "en_US"
-        break
-      case "fi":
-        language = "fi_FI"
-        break
-      case "sv":
-        language = "sv_SE"
-        break
-      default:
-        language = "fi_FI"
     }
 
     const instructions = (
@@ -114,7 +119,7 @@ export function completionInstructions({ knex }: ApiContext) {
         .select<any, CourseTranslation[]>("instructions")
         .from("course_translation")
         .where("course_id", course.id)
-        .where("language", language)
+        .where("language", languageMap[language] ?? "fi_FI")
     )[0]?.instructions
 
     if (instructions) {
@@ -123,12 +128,10 @@ export function completionInstructions({ knex }: ApiContext) {
       return res.status(404).json("")
     }
   }
-}
 
-export function completionTiers(ctx: ApiContext) {
-  return async function (req: Request<{ slug: string }>, res: Response) {
-    const { knex } = ctx
-    const getUserResult = await getUser(ctx)(req, res)
+  completionTiers = async (req: Request<{ slug: string }>, res: Response) => {
+    const { knex } = this.ctx
+    const getUserResult = await getUser(this.ctx)(req, res)
 
     if (getUserResult.isErr()) {
       return getUserResult.error
@@ -203,10 +206,8 @@ export function completionTiers(ctx: ApiContext) {
       return res.status(200).json({ tierData })
     }
   }
-}
 
-export function recheckCompletion(ctx: ApiContext) {
-  return async function (
+  recheckCompletion = async (
     req: Request<
       {},
       {},
@@ -218,9 +219,9 @@ export function recheckCompletion(ctx: ApiContext) {
       }
     >,
     res: Response,
-  ) {
-    const { prisma, logger, knex } = ctx
-    const adminRes = await requireAdmin(ctx)(req, res)
+  ) => {
+    const { prisma, logger, knex } = this.ctx
+    const adminRes = await requireAdmin(this.ctx)(req, res)
 
     if (adminRes !== true) {
       return adminRes
@@ -352,5 +353,87 @@ export function recheckCompletion(ctx: ApiContext) {
     return res
       .status(200)
       .json({ message: "No change", completion: existingCompletion })
+  }
+
+  registerCompletions = async (
+    req: Request<
+      {},
+      {},
+      {
+        completions: RegisterCompletionInput[]
+      }
+    >,
+    res: Response,
+  ) => {
+    const { prisma } = this.ctx
+    const organizationResult = await getOrganization(this.ctx)(req, res)
+
+    if (organizationResult.isErr()) {
+      return organizationResult.error
+    }
+
+    const org = organizationResult.value
+
+    const { completions } = req.body ?? {}
+
+    if (!completions) {
+      return res
+        .status(400)
+        .json({ message: "must provide completions in post message body" })
+    }
+
+    const registerCompletionSchema = yup.array().of(
+      yup
+        .object()
+        .shape({
+          completion_id: yup.string().min(32).max(36).required(),
+          student_number: yup.string().required(),
+        })
+        .required(),
+    )
+
+    try {
+      await registerCompletionSchema.validate(completions)
+    } catch (error) {
+      return res.status(403).json({ message: "malformed data", error })
+    }
+
+    const buildPromises = (data: RegisterCompletionInput[]) => {
+      return data.map(async (entry) => {
+        const { course_id, user_id } =
+          (await prisma.completion.findUnique({
+            where: { id: entry.completion_id },
+            select: { course_id: true, user_id: true },
+          })) ?? {}
+
+        if (!course_id || !user_id) {
+          // TODO/FIXME: we now fail silently if course/user not found
+          return Promise.resolve()
+        }
+
+        return prisma.completionRegistered.create({
+          data: {
+            completion: {
+              connect: { id: entry.completion_id },
+            },
+            organization: {
+              connect: { id: org?.id },
+            },
+            course: { connect: { id: course_id } },
+            real_student_number: entry.student_number,
+            registration_date: entry.registration_date ?? null,
+            user: { connect: { id: user_id } },
+          },
+        })
+      })
+    }
+
+    const queue = chunk(completions, 500)
+    for (const completionChunk of queue) {
+      const promises = buildPromises(completionChunk)
+      await Promise.all(promises)
+    }
+
+    return res.status(200).json({ message: "success" })
   }
 }
