+++ backend/bin/lib/logger.ts
@@ -2,7 +2,8 @@ import winston, { format } from "winston"
 import WinstonSentry from "winston-sentry-log"
 
 import { isProduction } from "../../config"
-import { Sentry } from "../../services/sentry"
+
+// import { Sentry } from "../../services/sentry"
 
 interface LoggerOptions {
   service: string
@@ -16,13 +17,14 @@ export default function logger({ service }: LoggerOptions) {
   const transports: winston.transport[] = [new winston.transports.Console()]
 
   if (isProduction) {
+    console.log("prod?")
     transports.push(
       new WinstonSentry({
         tags: {
           service,
         },
         level: "error",
-        sentryClient: Sentry,
+        sentryClient: require("../../services/sentry").Sentry,
         isClientInitialized: true,
         fingerprint: service,
       }),
