+++ backend/server.ts
@@ -40,6 +40,7 @@ const createExpressAppWithContext = ({
   app.use(graphqlUploadExpress())
   app.use(express.json())
   app.use("/api", apiRouter({ prisma, knex, logger }))
+
   return app
 }
